Refactor handlers into dedicated modules
