+++ bot/handlers/__init__.py
@@ -1,0 +1,4 @@
+from .start import start
+from .navigation import render_state, echo_handler
+
+__all__ = ["start", "render_state", "echo_handler"]
