+++ bot/handlers/start.py
@@ -1,0 +1,9 @@
+from telegram import Update
+from telegram.ext import ContextTypes
+
+from ..keyboards import main_menu
+
+async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    await update.message.reply_text(
+        "👋 مرحبًا بك في بوت أرشيف قسم الميكاترونكس.\nاختر من القائمة:",
+        reply_markup=main_menu)
